Extracted deeplinks, inapp and push models from optimove_flutter
